requirements.txt
asgiref==3.6.0
decouple==0.0.7
Django==4.1.4
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
